Feature/api gateway

+++ api-gateway/src/main/java/com/nitb/apigateway/dto/Vocabulary/response/VocabularyWordsPaginationResponseDto.java
@@ -14,7 +14,7 @@ import java.util.List;
 public class VocabularyWordsPaginationResponseDto {
     private List<VocabularyWordResponseDto> words;
 
-    private Integer totalItems;
+    private Long totalItems;
 
     private Integer totalPages;
 }
